requirements.txt
# Core Flask
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
python-dotenv==1.0.0

# File Processing Libraries
Pillow==10.0.1
Wand==0.6.13
python-magic==0.4.27
filetype==1.2.0

# FFmpeg Python Wrapper (Video/Audio conversion)
ffmpeg-python==0.2.0

# Document Processing
python-docx==0.8.11
openpyxl==3.1.2
PyPDF2==3.0.1
pandoc==2.3
python-pptx==0.6.21
xlrd==2.0.1
xlwt==1.3.0
pandas==2.1.4

# Archive Handling
patoolib==1.12.0
py7zr==0.20.6

# Background Processing
celery==5.3.4
redis==5.0.1

# Optional: Cloud Conversion APIs (for backup/advanced formats)
requests==2.31.0

# Utilities
humanize==4.8.0
click==8.1.7

# Production
gunicorn==21.2.0